Bump lib9c 1.7.2

+++ MarketService.Response/Interface/ISkillSchema.cs
@@ -11,7 +11,7 @@ public interface ISkillSchema
     public SkillCategory SkillCategory { get; set; }
     public int HitCount { get; set; }
     public int Cooldown { get; set; }
-    public int Power { get; set; }
+    public long Power { get; set; }
     public int StatPowerRatio { get; set; }
     public int Chance { get; set; }
     public StatType ReferencedStatType { get; set; }
